+++ packages/core/botpress/src/about.js
@@ -10,8 +10,8 @@ module.exports = projectLocation => {
       name: packageJson.name,
       version: packageJson.version,
       description: packageJson.description || '<no description>',
-      author: packageJson.author || '<no author>',
-      license: packageJson.license || 'AGPL-v3.0'
+      author: (typeof packageJson.author === 'object' ? packageJson.author.name : packageJson.author) || '<no author>',
+      license: packageJson.license || 'AGPL-3.0-only'
     }
   }